Using new functions from 'random' module from cwutils/lib
The main place where the functions are used is test_framework: there was a
TODO that I wanted to address.

One of recent commits has been changing this area, so I thought that I can go
one step further and fix the TODO.

Notice that none of the changes related to random are really affecting
production code - just test code is affected.

+++ src/cwutils/cw_cmdline.c
@@ -944,7 +944,7 @@ int cw_process_option(int opt, const char *optarg, cw_config_t *config)
 		break;
 
 	case 'D':
-		config->test_random_seed = atol(optarg);
+		config->test_random_seed = (uint32_t) atol(optarg);
 		break;
 
 	default: /* '?' */
